+++ ldm/modules/diffusionmodules/util.py
@@ -163,9 +163,10 @@ def timestep_embedding(timesteps, dim, max_period=10000, repeat_only=False):
     if not repeat_only:
         half = dim // 2
         freqs = torch.exp(
-            -math.log(max_period) * torch.arange(start=0, end=half, dtype=torch.float32, device=timesteps.device) / half
+            -math.log(max_period) * torch.arange(start=0, end=half,
+             device=timesteps.device) / half
         )
-        args = timesteps[:, None].float() * freqs[None]
+        args = timesteps[:, None] * freqs[None]
         embedding = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
         if dim % 2:
             embedding = torch.cat([embedding, torch.zeros_like(embedding[:, :1])], dim=-1)
@@ -216,7 +217,7 @@ class SiLU(nn.Module):
 
 class GroupNorm32(nn.GroupNorm):
     def forward(self, x):
-        return super().forward(x.float()).type(x.dtype)
+        return super().forward(x)
 
 def conv_nd(dims, *args, **kwargs):
     """
